feat(console): add common console instance for shared output functionality refactor(formatting): move cpm and console_panel functions to formatting.py to resolve circular imports fix(fallback_handler, llm): update imports to use cpm from formatting module instead of output module
