+++ ra_aid/console/common.py
@@ -1,0 +1,8 @@
+"""
+Common resources for console output functionality.
+"""
+
+from rich.console import Console
+
+# Create a single console instance to be shared across modules
+console = Console()
